Move get_expected_*_for_version to history.py
